requirements.txt
fastapi
uvicorn[standard]
modelscope
torch
torchvision
transformers
python-multipart
pillow
aiofiles
